Time: 380 ms (11.89%), Space: 109.7 MB (9.78%) - LeetHub

+++ 643-maximum-average-subarray-i/643-maximum-average-subarray-i.cpp
@@ -5,13 +5,13 @@ public:
         double currentSum=0;
         int n = nums.size();
         for(auto i=0; i<n; i++) {
-            currentSum = currentSum + nums[i];
+            currentSum = currentSum + (double)nums[i];
             if(i>(k-2)) {
                 currentAverage = currentSum / k;
                 if(currentAverage > max) {
                     max = currentAverage;
                 }
-                currentSum = currentSum - nums[i-k+1];
+                currentSum = currentSum - (double)nums[i-k+1];
                 
             }
         }
